+++ src/packages/preset/CHANGELOG.md
@@ -13,6 +13,30 @@ and this project adheres to [Semantic Versioning][semver].
 
 ---
 
+## [0.8.1] - 04-11-2023
+
+<br />
+
+### Improved
+
+- Improved aspect ratio plugin code
+
+<br /><br />
+
+---
+
+## [0.8.0] - 04-11-2023
+
+<br />
+
+### Added
+
+- #101 &nbsp;-&nbsp; Added columns
+
+<br /><br />
+
+---
+
 ## [0.7.2] - 23-09-2023
 
 <br />
@@ -232,6 +256,8 @@ and this project adheres to [Semantic Versioning][semver].
 
 शुभारम्भः
 
+[0.8.1]: https://github.com/mrjadeja/vishesh/commit/43dd2a53...19d2e1d6
+[0.8.0]: https://github.com/mrjadeja/vishesh/commit/e38f534e...43dd2a53
 [0.7.2]: https://github.com/mrjadeja/vishesh/commit/d45c96e9...e38f534e
 [0.7.1]: https://github.com/mrjadeja/vishesh/commit/de281dd1...d45c96e9
 [0.7.0]: https://github.com/mrjadeja/vishesh/commit/f709c3c1...de281dd1